+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 ## Changelog
+- New: Integrated Terminal
 - Fixed: HTML Preview, When you link css files in your HTML page
+- Improvement: App size
 - Improvement: UI